flowo-agno-service/pyproject.toml
[project]
name = "flowo-agno-service"
version = "0.1.0"
description = "Flowo AI Agent Service for flower recommendations and search"
requires-python = ">=3.12"
dependencies = [
    "agno>=0.1.0",
    "anthropic>=0.40.0",
    "openai>=1.55.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "copilotkit>=0.1.0",
    "lancedb>=0.17.0",
    "tantivy>=0.22.0",
    "sqlalchemy>=2.0.0",
    "pylance>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
]